Remove old ManageUserProfile
